UI Update 4: Confirm Quit Menu & Canvas Fix
I first fixed the issue regarding the canvas sizing and button placement. Your background size was way too large for the set canvas, so I adjusted that properly

Second, I added a "Quit Confirm Menu" that simply asks the user if they want to quit. This is triggered by clicking the Quit button. The "Yes" button now has the function of the quit button, and the "No" button returns to the main menu.