Enemy Update!
There is a GameObject that controls enemy spawns with a script, by randomly choosing 1 of 3 available spawn locations. Another script controls the movement, which is chosen at random.

Fixed a camera issue with sprites not appearing in the game tab. This was done by tweaking the Z coordinate values of some sprites and the camera position.